Stash the data
