+++ src/main/java/data_access/ProteinDataAccessFactory.java
@@ -1,7 +1,7 @@
 package data_access;
 
 public class DiseaseDataAccessFactory {
-    public DiseaseDataAccessObject create(String proteinName) {
+    public DiseaseDataAccessObject create(String proteinName) throws Exception {
         return new DiseaseDataAccessObject(proteinName);
     }
 }
